% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FOT_functionsOostveen.R
\name{compute_oostveen_r_percent_pred}
\alias{compute_oostveen_r_percent_pred}
\title{Compute Oostveen Resistance Percent Predicted}
\usage{
compute_oostveen_r_percent_pred(
  sex,
  ht,
  ht_unit = "cm",
  age,
  weight,
  measured,
  weight_unit = "kg",
  pressure_unit = "cmh2o",
  frequency,
  ...
)
}
\arguments{
\item{sex}{Sex as character string: "Male" or "Female"}

\item{ht}{Height (numeric)}

\item{ht_unit}{Height units: "cm" (default), "in", or "m"}

\item{age}{Age in years (numeric)}

\item{weight}{Body weight (numeric)}

\item{measured}{Measured resistance value (numeric)}

\item{weight_unit}{Weight units: "kg" (default) or "lbs"}

\item{pressure_unit}{Pressure units: "cmh2o" (default) or "hpa}

\item{frequency}{Frequency in Hz (character)}

\item{...}{Additional arguments (ignored)}
}
\value{
Percent predicted (numeric)
}
\description{
Calculate percent predicted for measured airway resistance.
}
\examples{
compute_oostveen_r_percent_pred(sex = "Male", ht = 180, age = 50, weight = 80,
                                measured = 0.35, frequency = "5")
}
\references{
Oostveen E, et al. Eur Respir J. 2013;42(6):1513-1523.
}
