% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FOT_functionsOostveen.R
\name{compute_oostveen_r_pred}
\alias{compute_oostveen_r_pred}
\title{Compute Oostveen Resistance Prediction}
\usage{
compute_oostveen_r_pred(
  sex,
  ht,
  ht_unit = "cm",
  age,
  weight,
  weight_unit = "kg",
  pressure_unit = "cmh2o",
  frequency,
  ...
)
}
\arguments{
\item{sex}{Sex as character string: "Male" or "Female"}

\item{ht}{Height (numeric)}

\item{ht_unit}{Height units: "cm" (default), "in", or "m"}

\item{age}{Age in years (numeric)}

\item{weight}{Body weight (numeric)}

\item{weight_unit}{Weight units: "kg" (default) or "lbs"}

\item{pressure_unit}{Pressure units: "cmh2o" (default) or "hpa}

\item{frequency}{Frequency in Hz: "4", "5", "6", "7", "8", "10", "11", "12", 
"14", "15", "16", "18", "19", "20", "22", "24", "26"}

\item{...}{Additional arguments (ignored)}
}
\value{
Predicted resistance in hPa·s·L⁻¹ or cmH₂O·s·L⁻¹
}
\description{
Calculate predicted airway resistance using Oostveen et al. 2013 equations.
}
\examples{
compute_oostveen_r_pred(sex = "Male", ht = 180, age = 50, weight = 80, frequency = "5")
}
\references{
Oostveen E, et al. The forced oscillation technique in clinical 
  practice: methodology, recommendations and future developments. 
  Eur Respir J. 2013;42(6):1513-1523.
}
