% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FOT_functionsOostveen.R
\name{compute_oostveen_x_pred}
\alias{compute_oostveen_x_pred}
\title{Compute Oostveen Reactance Prediction}
\usage{
compute_oostveen_x_pred(
  sex,
  ht,
  ht_unit = "cm",
  age,
  weight,
  weight_unit = "kg",
  pressure_unit = "cmh2o",
  frequency,
  ...
)
}
\arguments{
\item{sex}{Sex as character string: "Male" or "Female"}

\item{ht}{Height (numeric)}

\item{ht_unit}{Height units: "cm" (default), "in", or "m"}

\item{age}{Age in years (numeric)}

\item{weight}{Body weight (numeric)}

\item{weight_unit}{Weight units: "kg" (default) or "lbs"}

\item{pressure_unit}{Pressure units: "cmh2o" (default) or "hpa"  (not used for fres)}

\item{frequency}{Frequency: "4", "5", "6", "8", "10", "11", "12", "14", "fres", "AX4", "AX5"}

\item{...}{Additional arguments (ignored)}
}
\value{
Predicted reactance in hPa·s·L⁻¹ (or Hz for fres)
}
\description{
Calculate predicted reactance using Oostveen et al. 2013 equations.
}
\examples{
compute_oostveen_x_pred(sex = "Female", ht = 165, age = 55, weight = 65, frequency = "5")
}
\references{
Oostveen E, et al. Eur Respir J. 2013;42(6):1513-1523.
}
