% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gli_global_functions.R
\name{lookup_spline}
\alias{lookup_spline}
\title{Lookup GLI 2021 Spline Values}
\usage{
lookup_spline(age, sex, variable)
}
\arguments{
\item{age}{Numeric. Age in years}

\item{sex}{Character. Sex ("M", "Male", "MALE", "F", "Female", "FEMALE")}

\item{variable}{Character. Spirometric parameter ("FEV1", "FVC", "FEV1FVC")}
}
\value{
List with Mspline and Sspline values
}
\description{
Helper function to extract M spline and S spline values from lookup tables
based on age, sex, and spirometric parameter.
}
\keyword{internal}
