% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLI_dlco_functions.r
\name{lookup_spline_dlco}
\alias{lookup_spline_dlco}
\title{Lookup GLI DLCO Spline Values}
\usage{
lookup_spline_dlco(age, sex, variable)
}
\arguments{
\item{age}{Numeric. Age in years}

\item{sex}{Character. Sex ("M", "Male", "MALE", "F", "Female", "FEMALE")}

\item{variable}{Character. DLCO parameter ("DLCO", "KCO", "VA")}
}
\value{
List with Mspline and Sspline values
}
\description{
Helper function to extract M spline and S spline values from DLCO lookup tables
based on age, sex, and DLCO parameter.
}
\keyword{internal}
